[TASK] handling plane now function
