+++ src/bpm_data_combiner/app/controller.py
@@ -1,32 +1,22 @@
 from enum import Enum
-from typing import Sequence, Union
 import logging
+from typing import Sequence, Union
 
-from bpm_data_combiner.app.config import Config
-from bpm_data_combiner.monitor_devices.bl.monitor_synchronisation import (
-    MonitorDeviceSynchronisation, offset_from_median,
-)
-from bpm_data_combiner.monitor_devices.data_model.monitored_device import (
-    MonitoredDevice,
-)
+from collector import Collector, CollectionItemInterface
 
 from ..bl.accumulator import Accumulator
 from ..data_model.bpm_data_reading import BPMReading
 from ..interfaces.controller import ControllerInterface
-from ..monitor_devices.bl.monitor_devices_status import MonitorDevicesStatus
-from ..monitor_devices.interfaces.monitor_devices_status import (
-    MonitorDevicesStatusInterface,
-    StatusField,
-)
+from ..monitor_devices import MonitorDevicesStatus, MonitorDeviceSynchronisation, StatusField
 from ..post_processor.combine import collection_to_bpm_data_collection, accumulated_collections_to_array
-from ..post_processor.preprocessor import PreProcessor
+from ..post_processor.handle_active_planes import pass_data_for_active_planes
 from ..post_processor.statistics import compute_mean_weights_for_planes
 
+from .config import Config
 from .view import Views
 
-from collector import Collector, CollectionItemInterface
-
 logger = logging.getLogger("bpm-data-combiner")
+
 
 class ValidCommands(Enum):
     # Device data
@@ -40,7 +30,6 @@ class ValidCommands(Enum):
     periodic = "periodic"
     cfg_comp_median = "cfg_comp_median"
     known_device_names = "known_device_names"
-
 
 
 class Controller(ControllerInterface):
@@ -61,10 +50,6 @@ class Controller(ControllerInterface):
         )
 
         self.collector = Collector(devices_names=self.monitor_devices.get_device_names())
-        # Todo: make it a function and move it to on_collection_ready
-        self.preprocessor = PreProcessor(
-            devices_status=self.monitor_devices.devices_status
-        )
 
     def set_device_names(self, device_names=Sequence[str]):
         self.dev_name_index = {name: idx for idx, name in enumerate(device_names)}
@@ -120,10 +105,12 @@ class Controller(ControllerInterface):
 
     def new_value(self, dev_name: str, value: Sequence[int]):
         cnt, x, y = value
-        # when the one collection is ready it
-        # Todo: evaluate if a collection is ready
-        #       if so trigger self._on_new_collection_ready()
-        collection = self.collector.new_item(BPMReading(cnt=cnt, x=x, y=y, dev_name=dev_name))
+        collection = self.collector.new_item(
+            pass_data_for_active_planes(
+                cnt, x, y,
+                device_status=self.monitor_devices.devices_status[dev_name]
+            )
+        )
         if collection.ready:
             self._on_new_collection_ready(collection.data())
         if self.config.do_median_computation:
@@ -162,7 +149,6 @@ class Controller(ControllerInterface):
         # needs to know which are active
         self.collector.device_names = dev_names
         # needs to know which are active
-        self.preprocessor.device_names = dev_names
         devices = self.monitor_devices.devices_status
         self.views.monitor_bpms.update(
             names=[ds.name for _, ds in devices.items()],
