+++ src/tests/test_main.py
@@ -13,6 +13,7 @@ from known_devices import dev_names_bessyii as dev_names
 def test10_update_reset():
     update(dev_name=None, reset=True)
 
+
 def test20_reading():
     update(dev_name=None, reset=True)
     update(dev_name=dev_names[0], reading=np.array([42, 2**13, 2**14]).astype(np.int32))
